Remove need to subscribe to PreparingUpdateRequest
Update operations would fail if the user did not subscribe to the
SObject.PreparingUpdateRequest event … all they needed for success was
a subscribed handler, even if that handler did zero work. This change
removes the need to subscribe to the event. Now clients only need
subscribe if they have work to do in their handler.