Bug fixes: Auth Complete event and Refresh Token
This commit has two bug fixes.

1. AuthenticationComplete event
a) The previous implementation only raised the event on successful
authentication. This did not match the recommended style for client
code included in the samples: the samples recommend testing the
IsAuthenticated flag in the event args to determine success; however,
the event was only being raised on success so this test was
unnecessary. This commit changes the behavior so the event is raised
every time the underlying OAuth2Authenticator reports that
authentication has completed. The behavior will now match the
recommended style for the client code in the provided samples. The
behavior should now match user expectations.

b) ForceUserReauthorization(false) is now called upon successful
authentication. Previously, it was called only if the authentication
was successful AND the client has subscribed to the event. This was
unintuitive behavior: the SaleforceClient should not change when/how it
saves user accounts to disk based on whether the client subscribed a
completed handler.

c) The ‘sender’ argument passed to clients via this event was changed
from the OAuth2Authenticator to the SalesforceClient. Passing the
OAuth2Authenticator violated the .NET guidelines for events and it
exposed a protected field to the client.

2. Refresh token bug fix
The private method RefreshSessionToken was failing if the user had
logged in during the current session. This is because the
SalesforceClient constructor sets AccessTokenUrl to null when a login
is needed in order to force the OAuth2Authenticator to use the
User-Agent Flow rather than the Web-Server Flow. Since AccessTokenUrl
is null in the Authenticator, the WebRequest.Create() call in the
Authenticator throws an exception when this refresh code executes.
Refresh did work if the user logged in, saved the user to disk, closed
the app, and restarted it since then the SalesforceClient constructor
did not set AccessTokenUrl to null in the Authenticator. This bug fix
allows refresh to work in both cases.